+++ test/test.vcxproj
@@ -138,13 +138,10 @@
     <ClInclude Include="..\include\elastic\access.hpp" />
     <ClInclude Include="..\include\elastic\archive_exception.hpp" />
     <ClInclude Include="..\include\elastic\type.hpp" />
-    <ClInclude Include="..\include\elastic\basic_iprimitive.hpp" />
-    <ClInclude Include="..\include\elastic\basic_oprimitive.hpp" />
+    <ClInclude Include="..\include\elastic\binary_primitive.hpp" />
     <ClInclude Include="..\include\elastic\basic_primitive.hpp" />
     <ClInclude Include="..\include\elastic\binary_iarchive.hpp" />
     <ClInclude Include="..\include\elastic\binary_iserialize.hpp" />
-    <ClInclude Include="..\include\elastic\binary_oarchive.hpp" />
-    <ClInclude Include="..\include\elastic\binary_oserialize.hpp" />
     <ClInclude Include="..\include\elastic\concepts.hpp" />
     <ClInclude Include="..\include\elastic\interface_archive.hpp" />
     <ClInclude Include="..\include\elastic\property.hpp" />
